Update tests user ride offer
